src/backend/requirements.txt
flask==3.1.1
python-dotenv==1.1.0

langchain==0.3.26
langchain-openai==0.3.24
langchain-pinecone==0.2.8
langchain-community==0.3.26
langchain-google-genai==2.0.7

pinecone-client==3.2.2
cohere==5.6.2

pypdf==5.6.1
sentence-transformers==4.1.0

torch --extra-index-url https://download.pytorch.org/whl/cpu

requests==2.32.3
